# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Gerenciador de Tarefas - Sistema de Triagem e Atendimento
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type TaskPriority = typeof TaskPriority[keyof typeof TaskPriority];


export const TaskPriority = {
  critical: 'critical',
  high: 'high',
  medium: 'medium',
  low: 'low',
} as const;

export type TaskStatus = typeof TaskStatus[keyof typeof TaskStatus];


export const TaskStatus = {
  pending: 'pending',
  in_progress: 'in_progress',
  done: 'done',
  cancelled: 'cancelled',
} as const;

export type TaskSourceType = typeof TaskSourceType[keyof typeof TaskSourceType];


export const TaskSourceType = {
  manual: 'manual',
  text: 'text',
  pdf: 'pdf',
  spreadsheet: 'spreadsheet',
  audio: 'audio',
} as const;

export interface Task {
  id: number;
  title: string;
  /** @nullable */
  description?: string | null;
  priority: TaskPriority;
  status: TaskStatus;
  /** @nullable */
  categoryId?: number | null;
  /** @nullable */
  categoryName?: string | null;
  /** @nullable */
  categoryColor?: string | null;
  /** @nullable */
  assignee?: string | null;
  /** @nullable */
  dueDate?: string | null;
  sourceType?: TaskSourceType;
  /**
     * @minimum 0
     * @maximum 100
     */
  progress?: number;
  /** @nullable */
  startTime?: string | null;
  /** @nullable */
  endTime?: string | null;
  createdAt: string;
  updatedAt?: string;
}

export type TaskInputPriority = typeof TaskInputPriority[keyof typeof TaskInputPriority];


export const TaskInputPriority = {
  critical: 'critical',
  high: 'high',
  medium: 'medium',
  low: 'low',
} as const;

export type TaskInputStatus = typeof TaskInputStatus[keyof typeof TaskInputStatus];


export const TaskInputStatus = {
  pending: 'pending',
  in_progress: 'in_progress',
  done: 'done',
  cancelled: 'cancelled',
} as const;

export type TaskInputSourceType = typeof TaskInputSourceType[keyof typeof TaskInputSourceType];


export const TaskInputSourceType = {
  manual: 'manual',
  text: 'text',
  pdf: 'pdf',
  spreadsheet: 'spreadsheet',
  audio: 'audio',
} as const;

export interface TaskInput {
  /** @minLength 1 */
  title: string;
  description?: string;
  priority: TaskInputPriority;
  status?: TaskInputStatus;
  categoryId?: number;
  assignee?: string;
  dueDate?: string;
  sourceType?: TaskInputSourceType;
  /**
     * @minimum 0
     * @maximum 100
     */
  progress?: number;
}

export type TaskUpdatePriority = typeof TaskUpdatePriority[keyof typeof TaskUpdatePriority];


export const TaskUpdatePriority = {
  critical: 'critical',
  high: 'high',
  medium: 'medium',
  low: 'low',
} as const;

export type TaskUpdateStatus = typeof TaskUpdateStatus[keyof typeof TaskUpdateStatus];


export const TaskUpdateStatus = {
  pending: 'pending',
  in_progress: 'in_progress',
  done: 'done',
  cancelled: 'cancelled',
} as const;

export interface TaskUpdate {
  /** @minLength 1 */
  title?: string;
  description?: string;
  priority?: TaskUpdatePriority;
  status?: TaskUpdateStatus;
  /** @nullable */
  categoryId?: number | null;
  assignee?: string;
  /** @nullable */
  dueDate?: string | null;
  /**
     * @minimum 0
     * @maximum 100
     */
  progress?: number;
}

export type StatusUpdateStatus = typeof StatusUpdateStatus[keyof typeof StatusUpdateStatus];


export const StatusUpdateStatus = {
  pending: 'pending',
  in_progress: 'in_progress',
  done: 'done',
  cancelled: 'cancelled',
} as const;

export interface StatusUpdate {
  status: StatusUpdateStatus;
}

export interface Category {
  id: number;
  name: string;
  color: string;
  taskCount?: number;
}

export interface CategoryInput {
  /** @minLength 1 */
  name: string;
  color: string;
}

export type TaskStatsByStatus = {
  pending: number;
  in_progress: number;
  done: number;
  cancelled: number;
};

export type TaskStatsByPriority = {
  critical: number;
  high: number;
  medium: number;
  low: number;
};

export interface TaskStats {
  total: number;
  byStatus: TaskStatsByStatus;
  byPriority: TaskStatsByPriority;
  completionRate: number;
}

export type ImportInputSourceType = typeof ImportInputSourceType[keyof typeof ImportInputSourceType];


export const ImportInputSourceType = {
  text: 'text',
  pdf: 'pdf',
  spreadsheet: 'spreadsheet',
  audio: 'audio',
} as const;

export type ImportInputDefaultPriority = typeof ImportInputDefaultPriority[keyof typeof ImportInputDefaultPriority];


export const ImportInputDefaultPriority = {
  critical: 'critical',
  high: 'high',
  medium: 'medium',
  low: 'low',
} as const;

export interface ImportInput {
  /** Raw text content parsed from the uploaded file or transcribed audio */
  content: string;
  sourceType: ImportInputSourceType;
  defaultPriority?: ImportInputDefaultPriority;
  defaultCategoryId?: number;
}

export interface Subtask {
  id: number;
  taskId: number;
  title: string;
  done: boolean;
  position: number;
  createdAt: string;
}

export interface SubtaskInput {
  /** @minLength 1 */
  title: string;
  position?: number;
}

export interface SubtaskUpdate {
  /** @minLength 1 */
  title?: string;
  done?: boolean;
}

export interface ImportResult {
  imported: number;
  tasks: Task[];
}

export type ListTasksParams = {
status?: string;
priority?: string;
categoryId?: number;
search?: string;
};