# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Gerenciador de Tarefas - Sistema de Triagem e Atendimento
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: tasks
    description: Task management
  - name: categories
    description: Category management
  - name: dashboard
    description: Dashboard and stats

paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /tasks:
    get:
      operationId: listTasks
      tags: [tasks]
      summary: List all tasks
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: priority
          in: query
          schema:
            type: string
        - name: categoryId
          in: query
          schema:
            type: integer
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
    post:
      operationId: createTask
      tags: [tasks]
      summary: Create a new task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"

  /tasks/stats:
    get:
      operationId: getTaskStats
      tags: [dashboard]
      summary: Get task statistics for dashboard
      responses:
        "200":
          description: Task statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskStats"

  /tasks/recent:
    get:
      operationId: getRecentTasks
      tags: [dashboard]
      summary: Get recently updated tasks
      responses:
        "200":
          description: Recent tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"

  /tasks/import:
    post:
      operationId: importTasks
      tags: [tasks]
      summary: Import tasks from text content (parsed from PDF, spreadsheet, audio transcription, or plain text)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ImportInput"
      responses:
        "201":
          description: Imported tasks
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImportResult"

  /tasks/{id}:
    get:
      operationId: getTask
      tags: [tasks]
      summary: Get a single task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "404":
          description: Not found
    patch:
      operationId: updateTask
      tags: [tasks]
      summary: Update a task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskUpdate"
      responses:
        "200":
          description: Updated task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
    delete:
      operationId: deleteTask
      tags: [tasks]
      summary: Delete a task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /tasks/{id}/status:
    patch:
      operationId: updateTaskStatus
      tags: [tasks]
      summary: Update task status only (quick action)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StatusUpdate"
      responses:
        "200":
          description: Updated task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"

  /tasks/{id}/subtasks:
    get:
      operationId: listSubtasks
      tags: [tasks]
      summary: List subtasks for a task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of subtasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Subtask"
    post:
      operationId: createSubtask
      tags: [tasks]
      summary: Create a subtask
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubtaskInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subtask"

  /tasks/{id}/subtasks/{subtaskId}:
    patch:
      operationId: updateSubtask
      tags: [tasks]
      summary: Update a subtask
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: subtaskId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubtaskUpdate"
      responses:
        "200":
          description: Updated subtask
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subtask"
    delete:
      operationId: deleteSubtask
      tags: [tasks]
      summary: Delete a subtask
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: subtaskId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /categories:
    get:
      operationId: listCategories
      tags: [categories]
      summary: List all categories
      responses:
        "200":
          description: Categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
    post:
      operationId: createCategory
      tags: [categories]
      summary: Create a category
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"

  /categories/{id}:
    delete:
      operationId: deleteCategory
      tags: [categories]
      summary: Delete a category
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Task:
      type: object
      required:
        - id
        - title
        - priority
        - status
        - createdAt
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        priority:
          type: string
          enum: [critical, high, medium, low]
        status:
          type: string
          enum: [pending, in_progress, done, cancelled]
        categoryId:
          type: ["integer", "null"]
        categoryName:
          type: ["string", "null"]
        categoryColor:
          type: ["string", "null"]
        assignee:
          type: ["string", "null"]
        dueDate:
          type: ["string", "null"]
        sourceType:
          type: string
          enum: [manual, text, pdf, spreadsheet, audio]
        progress:
          type: integer
          minimum: 0
          maximum: 100
        startTime:
          type: ["string", "null"]
        endTime:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string

    TaskInput:
      type: object
      required:
        - title
        - priority
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: string
        priority:
          type: string
          enum: [critical, high, medium, low]
        status:
          type: string
          enum: [pending, in_progress, done, cancelled]
        categoryId:
          type: integer
        assignee:
          type: string
        dueDate:
          type: string
        sourceType:
          type: string
          enum: [manual, text, pdf, spreadsheet, audio]
        progress:
          type: integer
          minimum: 0
          maximum: 100

    TaskUpdate:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: string
        priority:
          type: string
          enum: [critical, high, medium, low]
        status:
          type: string
          enum: [pending, in_progress, done, cancelled]
        categoryId:
          type: ["integer", "null"]
        assignee:
          type: string
        dueDate:
          type: ["string", "null"]
        progress:
          type: integer
          minimum: 0
          maximum: 100

    StatusUpdate:
      type: object
      required:
        - status
      properties:
        status:
          type: string
          enum: [pending, in_progress, done, cancelled]

    Category:
      type: object
      required:
        - id
        - name
        - color
      properties:
        id:
          type: integer
        name:
          type: string
        color:
          type: string
        taskCount:
          type: integer

    CategoryInput:
      type: object
      required:
        - name
        - color
      properties:
        name:
          type: string
          minLength: 1
        color:
          type: string

    TaskStats:
      type: object
      required:
        - total
        - byStatus
        - byPriority
        - completionRate
      properties:
        total:
          type: integer
        byStatus:
          type: object
          properties:
            pending:
              type: integer
            in_progress:
              type: integer
            done:
              type: integer
            cancelled:
              type: integer
          required: [pending, in_progress, done, cancelled]
        byPriority:
          type: object
          properties:
            critical:
              type: integer
            high:
              type: integer
            medium:
              type: integer
            low:
              type: integer
          required: [critical, high, medium, low]
        completionRate:
          type: number

    ImportInput:
      type: object
      required:
        - content
        - sourceType
      properties:
        content:
          type: string
          description: Raw text content parsed from the uploaded file or transcribed audio
        sourceType:
          type: string
          enum: [text, pdf, spreadsheet, audio]
        defaultPriority:
          type: string
          enum: [critical, high, medium, low]
        defaultCategoryId:
          type: integer

    Subtask:
      type: object
      required:
        - id
        - taskId
        - title
        - done
        - position
        - createdAt
      properties:
        id:
          type: integer
        taskId:
          type: integer
        title:
          type: string
        done:
          type: boolean
        position:
          type: integer
        createdAt:
          type: string

    SubtaskInput:
      type: object
      required:
        - title
      properties:
        title:
          type: string
          minLength: 1
        position:
          type: integer

    SubtaskUpdate:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        done:
          type: boolean

    ImportResult:
      type: object
      required:
        - imported
        - tasks
      properties:
        imported:
          type: integer
        tasks:
          type: array
          items:
            $ref: "#/components/schemas/Task"