# lib/api-zod/src/generated/types/subtask.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Gerenciador de Tarefas - Sistema de Triagem e Atendimento
 * OpenAPI spec version: 0.1.0
 */

export interface Subtask {
  id: number;
  taskId: number;
  title: string;
  done: boolean;
  position: number;
  createdAt: string;
}