# lib/api-zod/src/generated/types/subtaskUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Gerenciador de Tarefas - Sistema de Triagem e Atendimento
 * OpenAPI spec version: 0.1.0
 */

export interface SubtaskUpdate {
  /** @minLength 1 */
  title?: string;
  done?: boolean;
}